+++ cpp/bisGridTransformation.h
@@ -126,7 +126,7 @@ class bisGridTransformation : public bisAbstractTransformation {
   virtual void serializeInPlace(unsigned char* output);
 
   /** returns size needed to serialize this object in bytes */
-  virtual int getRawSize();
+  virtual long getRawSize();
 
   /** returns raw data as float pointer */
   virtual float* getData() { return displacementField->getData(); }
